requirements.txt
pillow
python-dotenv
pypng==0.20220715.0
typing_extensions==4.10.0
qrcode[pil]
validators